Common: moved all data upgrading from GameSetupStruct to MainGameFile
