+++ Common/ac/gamesetupstruct.h
@@ -83,8 +83,6 @@ struct GameSetupStruct: public GameSetupStructBase {
     // I also had to move BuildAudioClipArray from the engine and make it
     // GameSetupStruct member.
 
-    void BuildAudioClipArray(const AGS::Common::AssetLibInfo &lib);
-
     //--------------------------------------------------------------------
     // Do not call these directly
     //------------------------------
@@ -125,11 +123,9 @@ struct GameSetupStruct: public GameSetupStructBase {
 
 //=============================================================================
 
-// Game conversion, fixup and compatibility support functions
+// TODO: find out how this function was supposed to be used
 void ConvertOldGameStruct (OldGameSetupStruct *ogss, GameSetupStruct *gss);
 // Finds an audio clip using legacy convention index
 ScriptAudioClip* GetAudioClipForOldStyleNumber(GameSetupStruct &game, bool is_music, int num);
-// Ensures that the game saves directory path is valid
-void FixupSaveDirectory(GameSetupStruct &game);
 
 #endif // __AGS_CN_AC__GAMESETUPSTRUCT_H
